fast/regex/regexp-cross-frame-callable-expected.txt

PASS successfullyParsed is true

TEST COMPLETE
PASS re('a') is ['a']
